Move Keycloak integration code to its own module
Keycloak integration is not related to the connected services.
